requirements.txt
google-cloud-aiplatform[langchain,reasoningengine]
cloudpickle==3.0.0
pydantic==2.7.4
langgraph
httpx
python-dotenv
setuptools